--- public/assets/glossary.ts
@@ -1,26 +1,0 @@
-export const glossary = [
-	{
-		term: "Alphabetical order glossary of terms",
-		definition: "With definitions provided by queer resources and organizations"
-	},
-	{
-		term: "Include any submitted identifiers by participants",
-		definition: "Even if a label is less known or more obscure, each individual that contributes should have their identity recognized if accurate definitions are possible to obtain"
-	},
-	{
-		term: "Etc..:",
-		definition: ""
-	},
-	{
-		term: "Etc..:",
-		definition: ""
-	},
-	{
-		term: "Etc..:",
-		definition: ""
-	},
-	{
-		term: "Etc..:",
-		definition: ""
-	},
-]
